Fix CSV tests on linux (stol not always 32bit)

+++ dropdbase/CSVDataImporter.cpp
@@ -303,7 +303,7 @@ DataType CSVDataImporter::IdentifyDataType(std::vector<std::string> columnValues
 		// COLUMN_INT
 		try {
 			size_t position;
-			std::stol(s, &position);
+			std::stoi(s, &position);
 			if (s.length() == position) {
 				dataTypes.push_back(COLUMN_INT);
 				continue;
